+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ignite-ui-lucariozin/tokens
 
+## 1.2.0
+
+### Minor Changes
+
+- Adding "Toast" component to "react" package
+
 ## 1.1.0
 
 ### Minor Changes